# swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "version":"1.0.0",
        "description":"API for E-Commerce application",
        "title":"E-commerce API"
    },
    "servers":[
        {
            "url":"http://localhost:3200"
        }, 
        {
            "url":"https://e-com-api-a29f.onrender.com"
        }
    ],

    "components":
    {
        "securitySchemes":
        {
            "JWT":
            {
                "in":"header",
                "name":"Authorization",
                "type": "apiKey"
            }
        }
    },

    "paths":
    {  
        "/api/users/signup": 
        {
            "post":
            {
                "tags": ["Users"],
                "summary":"SignUp",
                "description":"User SignUp",
                "requestBody":
                {
                    "content":
                    {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name":{
                                        "type": "string"
                                    },
                                    "email":{
                                        "type": "string"
                                    },
                                    "password":{
                                        "type": "string"
                                    },
                                    "type":{
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name":{
                                            "type": "string"
                                        },
                                        "email":{
                                            "type": "string"
                                        },
                                        "type":{
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect Credentials"
                    }
                }
            }
        }
        ,
        "/api/users/signin": 
        {
            "post":
            {
                "tags": ["Users"],
                "summary":"Login",
                "description":"User login to get token",
                "requestBody":
                {
                    "required": "true",
                    "content":
                    {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email":{
                                        "type": "string"
                                    },
                                    "password":{
                                        "type": "string"
                                    }
                                },
                                "required": ["email", "password"]
                            },
                            "example": {
                              "email": "user@example.com",
                              "password": "userpassword123"
                            }
                        }
                    }
                },
                "responses":{
                    "200": {
                        "description":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NjZlZGI3ODU2MWJkNjNmYTczMzM1ZDUiLCJlbWFpbCI6InNhY2hpbkt1bWFyQGdtYWlsLmNvbSIsImlhdCI6MTcxODU0MTI4MiwiZXhwIjoxNzE4NTQ0ODgyfQ.5jdmkQwKOgYRas8PUi44tw0c4RbmKecvo5NhQA92fN4"
                    },
                    "400": {
                        "description": "Incorrect Credentials"
                    }
                }
            }
        },
        "/api/products": {
            "get":
            {
                "tags": ["Products"],
                "summary":"Get All Products",
                "description":"User will get all products",
                "security":[{"JWT": {} }],
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "_id":{
                                                "type": "string"
                                            },
                                            "name":{
                                                "type": "string"
                                            },
                                            "desc":{
                                                "type": "string"
                                            },
                                            "price":{
                                                "type": "integer"
                                            },
                                            "imageUrl":{
                                                "type": "string"
                                            },
                                            "category":{
                                                "type": "string"
                                            },
                                            "sizes":{
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "stock":{
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post":
            {
                "tags": ["Products"],
                "summary":"Add Product",
                "description":"User will add the new product",
                "security":[{"JWT": {} }],
                "requestBody":
                {
                    "required": "true",
                    "content":
                    {
                        "multipart/form-data":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name":{
                                        "type": "string"
                                    },
                                    "desc":{
                                        "type": "string"
                                    },
                                    "price":{
                                        "type": "integer"
                                    },
                                    "imageUrl": {
                                      "type": "string",
                                      "format": "binary",
                                      "description": "Image file"
                                    },
                                    "category":{
                                        "type": "string"
                                    },
                                    "sizes":{
                                        "type": "string"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    }
                                },
                                "required": ["name", "desc", "price", "imageUrl", "category", "sizes", "stock"]
                            }
                        }
                    }
                },
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id":{
                                            "type": "string"
                                        },
                                        "name":{
                                            "type": "string"
                                        },
                                        "desc":{
                                            "type": "string"
                                        },
                                        "price":{
                                            "type": "integer"
                                        },
                                        "imageUrl":{
                                            "type": "string"
                                        },
                                        "category":{
                                            "type": "string"
                                        },
                                        "sizes":{
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "stock":{
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/products/{productId}": {
            "get":
            {
                "tags": ["Products"],
                "summary":"Get One Product",
                "description":"User will get one product using id",
                "security":[{"JWT": {} }],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": "true",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id":{
                                            "type": "string"
                                        },
                                        "name":{
                                            "type": "string"
                                        },
                                        "desc":{
                                            "type": "string"
                                        },
                                        "price":{
                                            "type": "integer"
                                        },
                                        "imageUrl":{
                                            "type": "string"
                                        },
                                        "category":{
                                            "type": "string"
                                        },
                                        "sizes":{
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "stock":{
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/products/filter": {
            "get":
            {
                "tags": ["Products"],
                "summary":"Filter Product",
                "description":"User will get filtered product based on minimum price, maximum price and category",
                "security":[{"JWT": {} }],
                "parameters": [
                    {
                        "name": "minPrice",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "maxPrice",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {                                        
                                            "name":{
                                                "type": "string"
                                            },
                                            "price":{
                                                "type": "integer"
                                            },
                                            "category":{
                                                "type": "string"
                                            },
                                            "ratings":{
                                                "type": "array",
                                                "items": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/products/rate": {
            "post":
            {
                "tags": ["Products"],
                "summary":"Rate Product",
                "description":"User will rate product using the product id",
                "security":[{"JWT": {} }],
                "requestBody":
                {
                    "required": "true",
                    "content":
                    {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productId":{
                                        "type": "string"
                                    },
                                    "rating":{
                                        "type": "number"
                                    }
                                },
                                "required": ["productId", "rating"]
                            },
                            "example": {
                              "productId": "6666f812f585696b12d83cd8",
                              "rating": "3"
                            }
                        }
                    }
                },
                "responses":
                {
                    "200": {
                        "description": "Rating has been added"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/products/averagePrice": {
            "get":
            {
                "tags": ["Products"],
                "summary":"Average Price",
                "description":"User will get the average price based on category",
                "security":[{"JWT": {} }],
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "_id":{
                                                "type": "string",
                                                "description": "The category ID"
                                            },
                                            "averagePrice":{
                                                "type": "number",
                                                "format": "float",
                                                "description": "The average price of products in the category"
                                            }
                                        }
                                    }
                                },
                                "example": { 
                                    "_id": "Cat1",
                                    "averagePrice": 9000
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/cartItems": {
            "get":
            {
                "tags": ["Cart Items"],
                "summary":"Get Cart Items",
                "description":"User will get all items added in the cart",
                "security":[{"JWT": {} }],
                "responses":
                {
                    "200": {
                        "content":
                        {
                            "application/json":{
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "_id":{
                                                "type": "string"
                                            },
                                            "productId":{
                                                "type": "string"
                                            },
                                            "quantity":{
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post":
            {
                "tags": ["Cart Items"],
                "summary":"Add Product in Cart",
                "description":"User will add the product in cart",
                "security":[{"JWT": {} }],
                "requestBody":
                {
                    "required": "true",
                    "content":
                    {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productId":{
                                        "type": "string"
                                    },
                                    "quantity":{
                                        "type": "integer"
                                    }
                                },
                                "required": ["productId", "quantity"]
                            }
                        }
                    }
                },
                "responses":
                {
                    "201": {
                        "description": "Cart is updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/cartItems/{cartItemId}": {
            "delete": 
            {
                "tags": ["Cart Items"],
                "summary":"Delete Cart Item",
                "description":"User will one product from cart items",
                "security":[{"JWT": {} }],
                "parameters": [
                    {
                        "name": "cartItemId",
                        "in": "path",
                        "required": "true",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses":
                {
                    "200": {
                        "description": "Cart Item is removed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            }
        }
    }
}